# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorCode": {
        "enum": [
          "AUTHENTICATION_REQUIRED",
          "AUTHENTICATION_INVALID",
          "ACTION_NOT_AUTHORIZED",
          "RESOURCE_NOT_FOUND",
          "EVIDENCE_AUDIO_TOO_LARGE",
          "STALE_OPERATION_VERSION",
          "STALE_DRAFT_VERSION",
          "IDEMPOTENCY_KEY_REUSED",
          "STATE_CONFLICT",
          "MANDATE_CONFLICT",
          "VALIDATION_ERROR",
          "RATE_LIMITED",
          "TELEPHONY_UNAVAILABLE",
          "TELEPHONY_OUTCOME_UNCERTAIN",
          "REALTIME_UNAVAILABLE",
          "INTERNAL_ERROR",
          "CONTRACT_NOT_IMPLEMENTED"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "current_draft_version": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Draft Version"
          },
          "current_operation_version": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Operation Version"
          },
          "field_issues": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Issues"
          },
          "message": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CreateOutboundCallRequest": {
        "additionalProperties": false,
        "properties": {
          "ai_disclosure_required": {
            "const": true,
            "title": "Ai Disclosure Required",
            "type": "boolean"
          },
          "authorized_at": {
            "format": "date-time",
            "title": "Authorized At",
            "type": "string"
          },
          "authorized_by": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Authorized By",
            "type": "string"
          },
          "call_session_id": {
            "format": "uuid",
            "title": "Call Session Id",
            "type": "string"
          },
          "destination_label": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Destination Label",
            "type": "string"
          },
          "recording_consent_required": {
            "const": false,
            "title": "Recording Consent Required",
            "type": "boolean"
          },
          "recording_mode": {
            "const": "DISABLED",
            "title": "Recording Mode",
            "type": "string"
          }
        },
        "required": [
          "call_session_id",
          "destination_label",
          "authorized_by",
          "authorized_at",
          "ai_disclosure_required",
          "recording_mode",
          "recording_consent_required"
        ],
        "title": "CreateOutboundCallRequest",
        "type": "object"
      },
...
      "FieldIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "field": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Field",
            "type": "string"
          },
          "message": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldIssue",
        "type": "object"
      },
...
      "OutboundCallResponse": {
        "additionalProperties": false,
        "properties": {
          "call_session_id": {
            "format": "uuid",
            "title": "Call Session Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "provider_call_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Provider Call Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "QUEUED",
              "INITIATED",
              "RINGING",
              "IN_PROGRESS",
              "COMPLETED",
              "BUSY",
              "FAILED",
              "NO_ANSWER",
              "CANCELED"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_updated_at": {
            "format": "date-time",
            "title": "Status Updated At",
            "type": "string"
          }
        },
        "required": [
          "call_session_id",
          "provider_call_id",
          "status",
          "created_at",
          "status_updated_at"
        ],
        "title": "OutboundCallResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "opaque demo token",
        "description": "Configured demo bearer token. No token value is published.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Thin HTTP contract boundary for the Volta hackathon demo.",
    "title": "Volta API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/operations/{operation_id}/outbound-calls": {
      "post": {
        "operationId": "create_outbound_call",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          },
          {
            "description": "Printable ASCII key reused for one logical mutation.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Printable ASCII key reused for one logical mutation.",
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[\\x20-\\x7E]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutboundCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundCallResponse"
                }
              }
            },
            "description": "The call was accepted."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication is missing or invalid.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The authenticated actor lacks authority.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The referenced resource was not found.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The request conflicts with current state or idempotency history.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The request does not satisfy the public contract.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The configured demo traffic boundary was exceeded.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the authorized mutation window can accept traffic.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              },
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "An unexpected failure was translated safely.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier preserved or assigned by the API.",
                "schema": {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Provider failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unavailable or uncertain outcome."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Provider timeout."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Outbound Call",
        "tags": [
          "telephony"
        ]
      }
    },
...
  }
}